+++ client/src/components/LogoutButton.tsx
@@ -1,0 +1,11 @@
+import { useCallback, ReactNode } from 'react'
+import { authService } from '@oneblink/apps'
+
+export function LogoutButton(children?: ReactNode) {
+  const logOut = useCallback(async () => {
+    authService.logoutHostedUI()
+    await authService.logout()
+  }, [])
+
+  return <button onClick={logOut}>{children || 'Log Out'}</button>
+}