+++ CHANGELOG.md
@@ -13,3 +13,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - definition service
 - referenceDataStore class
 - AutoSaveControlledForm component
+- Login and Logout button components